# server/app/api/swagger.yml
openapi: "3.0.3"
info:
  description: "This is the Databus API documentation."
  version: "2.1.0"
  title: "Databus API"
components:
  securitySchemes:
    ApiKeyAuth:        
      type: apiKey
      in: header       
      name: X-API-KEY  
paths:
...
  /{account}/collections/{id}:
    get:
      summary: "Get Collection"
      description: "Retrieve a Databus Collection."
      operationId: "get-collection"
      tags:
        - Collection
      parameters:
        - name: account
          in: path
          description: account namespace
          schema:
            type: string
          required: true
          example: '%USERNAME%'
        - name: id
          in: path
          description: the collection identifier
          schema:
            type: string
          required: true
          example: "yourCollection"
      responses:
        "200":
          description: "OK"
          content:
            text/sparql:
              schema:
                type: string
                example: "PREFIX rdfs:   <http://www.w3.org/2000/01/rdf-schema#>
                  PREFIX rdf:    <http://www.w3.org/1999/02/22-rdf-syntax-ns#>
                  PREFIX dcat:   <http://www.w3.org/ns/dcat#>
                  PREFIX dct:    <http://purl.org/dc/terms/>
                  PREFIX dcv: <https://dataid.dbpedia.org/databus-cv#>
                  PREFIX databus: <https://dataid.dbpedia.org/databus#>
                  SELECT ?file WHERE
                  {
                  	{
                  		GRAPH ?g
                  		{
                  			{
                  				?dataset databus:group <%DATABUS_RESOURCE_BASE_URL%/%USERNAME%/yourCollection> .
                  				{
                  					?dataset databus:artifact <%DATABUS_RESOURCE_BASE_URL%/%USERNAME%/yourCollection/first> .
                  				}
                  			}
                  			?dataset dcat:distribution ?distribution .
                  			?distribution databus:file ?file .
                  		}
                  	}
                  }"
            application/ld+json:
              schema:
                type: object
                example: '{
                  "@id": "%DATABUS_RESOURCE_BASE_URL%/%USERNAME%/%GROUP%",
                  "@type": "databus:Group",
                  "description": "%GROUP% datasets from DBpedia",
                  "title": "%GROUP% Data",
                  "@context": {
                    "title": {
                      "@id": "http://purl.org/dc/terms/title"
                    },
                    "description": {
                      "@id": "http://purl.org/dc/terms/description"
                    },
                    "dbo": "http://dbpedia.org/ontology/",
                    "sec": "https://w3id.org/security#",
                    "dataid": "http://dataid.dbpedia.org/ns/core#",
                    "dct": "http://purl.org/dc/terms/",
                    "dcv": "https://dataid.dbpedia.org/databus-cv#",
                    "xsd": "http://www.w3.org/2001/XMLSchema#",
                    "rdfs": "http://www.w3.org/2000/01/rdf-schema#",
                    "cert": "http://www.w3.org/ns/auth/cert#",
                    "dcat": "http://www.w3.org/ns/dcat#",
                    "foaf": "http://xmlns.com/foaf/0.1/"
                  }
                }'
        "404":
          description: "Not Found."
          content:
            text/plain:
              schema:
                type: string
    delete:
      summary: "Delete Collection"
      description: ""
      operationId: "delete-collection"
      tags:
        - Collection
      security:
        - ApiKeyAuth: []
      parameters:
        - name: account
          in: path
          description: account namespace
          schema:
            type: string
          required: true
          example: '%USERNAME%'
        - name: id
          in: path
          description: the collection identifier
          schema:
            type: string
          required: true
          example: "%GROUP%-%ARTIFACT%"
      responses:
        "200":
          description: "OK"
        "404":
          description: "Not Found."
  /api/collection/md5hash:
    get:
      summary: "Get Collection md5-Hash"
      description: "Get the md5-Hash of the collection contents. Can be used to check whether the selected files of a collection changed."
      operationId: "collection-md5"
      tags:
        - Collection
      parameters:
        - name: uri
          in: query
          description: collection uri
          schema:
            type: string
          required: true
          example: '%DATABUS_RESOURCE_BASE_URL%/%USERNAME%/collections/%GROUP%-%ARTIFACT%'
      responses:
        "200":
          description: "OK"
          content:
            string:
              schema:
                type: string
                format: binary
                example: '37dac6d1e315917bceba4f38564dc529'
...